+++ tests/pna/graph/conftest.py
@@ -11,7 +11,9 @@ from pixelator.pna.config import PNAAntibodyPanel
 
 
 @pytest.fixture(name="mock_panel")
-def mock_panel_fixture():
+def mock_panel_fixture(request):
+    version = getattr(request, "param", "0.0.0")
+
     mock_antibody_panel = create_autospec(PNAAntibodyPanel)
     mock_antibody_panel.markers = ["MarkerA", "MarkerB", "MarkerC"]
     # Each marker is duplicated in the panel,
@@ -31,8 +33,12 @@ def mock_panel_fixture():
             "conj_id": ["C001", "C002", "C003"],
         }
     )
+
+    if version.startswith("2"):
+        mock_antibody_panel.df.drop(columns=["nuclear"], inplace=True)
+
     mock_antibody_panel.name = "mock-panel"
-    mock_antibody_panel.version = "0.0.0"
+    mock_antibody_panel.version = version
     mock_antibody_panel.description = "Dummy panel data"
     mock_antibody_panel.aliases = ["mock_alias"]
     return mock_antibody_panel